backend/python/requirements.txt
numpy
pandas
google-generativeai
matplotlib
seaborn
xhtml2pdf